+++ core/src/test/scala/org/apache/spark/sql/expression/index/UnsignedTestSuite.scala
@@ -98,7 +98,7 @@ class UnsignedTestSuite extends BaseTiSparkTest {
       List.empty,
       List(List(1), List(2), List(3)),
       List(List(0), List(1), List(2), List(3), List(unsignedLongMaxValue)))
-    assert(queries.size == answers.size)
+    assert(queries.lengthCompare(answers.length) == 0)
     for (i <- queries.indices) {
       println(queries(i))
       explainAndRunTest(queries(i), rJDBC = answers(i), skipTiDB = true)